requirements.txt
PyYAML
txzmq
pycrypto
pyasn1
